refactor code (seperate utils and hooks)
