+++ gatsby-theme-oi-wiki/src/components/Link/ToolCard.tsx
@@ -9,37 +9,8 @@ import {
 } from '@material-ui/core'
 import { Alert } from '@material-ui/lab'
 import { PreviewData, FetchStatus } from './LinkTooltip'
+import {useDelay} from './hooks'
 
-function useDelay (onOpen: () => void, onClose: () => void, openDelay: number, closeDelay: number) : [() => void, () => void] {
-  const closeHandle = useRef(null), openHandle = useRef(null)
-  function open () : void {
-    if (closeHandle.current) { // 正在准备close，则不让它close
-      clearTimeout(closeHandle.current)
-      closeHandle.current = null
-    }
-    if (!openHandle.current) { // 如果之前没有 open 时间就创建一个
-      openHandle.current = setTimeout(() => {
-        onOpen()
-        openHandle.current = null
-      }, openDelay)
-    }
-  }
-  function close () : void {
-    if (closeHandle) { // 之前的 close 事件需要被清除
-      clearTimeout(closeHandle.current)
-      closeHandle.current = null
-    }
-    if(openHandle.current){ // 鼠标快进快出，则不显示
-      clearTimeout(openHandle.current)
-      openHandle.current = null
-    }
-    closeHandle.current = setTimeout(() => {
-      onClose()
-      closeHandle.current = null
-    }, closeDelay)
-  }
-  return [open, close]
-}
 
 type PositionAndSize = {
   pos: Position,
